financial-etl-framework/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "financial-etl-framework"
version = "1.0.0"
description = "Framework ETL para Data Warehouse financeiro com PostgreSQL"
authors = [
    {name = "Giovanni Muller", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Private"}
keywords = ["etl", "postgresql", "data-warehouse", "finance", "bigquery"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    "python-dateutil>=2.8.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "isort>=5.12.0",
    "pylint>=3.0.0",
    "bandit>=1.7.0",
    "pre-commit>=3.0.0",
    "httpx>=0.24.0",  # Para testes de FastAPI
]
notebooks = [
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
    "notebook>=7.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "bokeh>=3.0.0",
]
excel = [
    "openpyxl>=3.1.0",
    "xlrd>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/GSMuller/financial-etl-framework"
Repository = "https://github.com/GSMuller/financial-etl-framework"
Issues = "https://github.com/GSMuller/financial-etl-framework/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["financial_etl*"]
exclude = ["tests*", "docs*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--showlocals",
    "--tb=short",
    "--cov=financial_etl",
    "--cov-report=html",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=json",
    "--cov-branch",
    "--cov-fail-under=75",
    "--maxfail=5",
    "--durations=10",
]
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "api: marks tests as API tests",
    "services: marks tests for services",
]
filterwarnings = [
    "error",
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["*/__pycache__/*", "*/venv/*", "*/build/*"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
check_untyped_defs = true
exclude = [
    "venv",
    "build",
    "dist",
    "tests"
]

[tool.pylint.main]
max-line-length = 100
good-names = ["i", "j", "k", "ex", "Run", "_", "id", "db", "conn", "cursor"]
disable = [
    "C0111",  # missing-docstring
    "R0903",  # too-few-public-methods
    "C0103",  # invalid-name
]

[tool.pylint.format]
max-line-length = 100

[tool.bandit]
exclude_dirs = ["tests", "venv", "build"]
skips = ["B101", "B601"]

[tool.coverage.run]
source = ["src/financial_etl"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/site-packages/*",
    "*/venv/*",
]
branch = true

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
fail_under = 75

[tool.coverage.html]
directory = "htmlcov"
